Rename sidebar-workspace-archive-redirect.pure.ts to workspace-archive-redirect.ts
